Added Access Control

+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,15 @@ use App\Post;
 class PostController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -75,6 +84,10 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        //Check for correct user
+        if (auth()->user()->id !==$post->user_id) {
+            return redirect('/post')->with('error', 'Unauthorized Page');
+        }
         return view('post.edit')->with('post',$post);
     }
 
@@ -104,6 +117,10 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //Check for correct user
+        if (auth()->user()->id !==$post->user_id) {
+            return redirect('/post')->with('error', 'Unauthorized Page');
+        }
         $post->delete();
 
         return redirect('/post')->with('success', 'Deleted');
